payment-service/pyproject.toml
[tool.poetry]
name = "payment-service"
version = "0.1.0"
description = ""
authors = ["hamzah-syed <hamzah.syed17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
sqlmodel = "^0.0.21"
kafka-python = "^2.0.2"
pydantic = "^2.8.2"
aiokafka = "^0.11.0"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
stripe = "^10.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
